requirements_prod.txt
ipdb
numpy
pandas
matplotlib
scikit-image
opencv-python


# api
fastapi
python-multipart
uvicorn

# prevent bq 3+ db-dtypes package error when running training from bq
google-cloud-bigquery<3.0.0
google-cloud-storage
pyarrow
colorama

split-folders

pytest-shutil
